+++ src/components/Calendar.js
@@ -2,6 +2,8 @@ import FullCalendar from '@fullcalendar/react';
 
 import dayGridPlugin from '@fullcalendar/daygrid';
 import React, { useState } from 'react';
+
+import { useLocation } from 'react-router-dom';
 import { Component } from 'react';
 import interactionPlugin from "@fullcalendar/interaction";
 import productData from "../data/product-data.json";
@@ -26,6 +28,29 @@ export const StyleWrapper = styled.div`
   background: #FFFAF0;
 }
 `
+
+export const Main=()=>{ //시작페이지
+  return(
+    <div>
+      <h3>TEAM : 2 0 1 8</h3>
+    <h2>C A L E N D A R</h2>
+    </div>
+  );
+}
+
+
+export const Error=()=>{ //에러페이지
+    const locations = useLocation();
+    return(
+      <div>
+        <h3>Page not found at {locations.pathname}</h3>
+      </div>
+    );
+  }
+
+
+
+
 
 const calendar_data = firestore.collection("calendar_data");
 /*
@@ -52,9 +77,11 @@ const Calendar = (props) => {
     setVisible(false);
   };
 
-
   const handleDateClick = (arg) => { // bind with an arrow function
     console.log(arg)
+      const calendar_data = firestore.collection("calendar_data");
+      var event = prompt("일정을 입력하세요.",);
+      calendar_data.doc(arg.dateStr).set( { date : arg.dateStr , title : `${event}`})
   }
 
   // 클릭 시 이벤트 정보 받아옴
@@ -63,17 +90,25 @@ const Calendar = (props) => {
     }
 
   return (
+    <StyleWrapper>
       <FullCalendar
         plugins={[dayGridPlugin,interactionPlugin]}
-        initialView="dayGridMonth"
+        headerToolbar={{
+          left: "prev,next today",
+          center: "title",
+          right: "dayGridMonth,dayGridWeek,dayGridDay",
+        }}
         selectable = {true}
         dateClick={handleDateClick}
         eventClick={handleEventClick}
         select={handleDateSelect}
+        editable={true}
+        droppable={true}
         weekends={true}
-        events={[]}
-        height="90vh"
+        events={productData}
+        
       />
+      </StyleWrapper>
   );
 };
